refactor(core): move language and version options schemas into schemas/
