+++ packages/core/src/bible-versions.ts
@@ -1,6 +1,6 @@
 import * as z from 'zod/mini';
 import type { ApiClient } from './client';
-import { BibleVersionSchema } from './schemas';
+import { GetVersionsOptionsSchema, type GetVersionsOptions } from './schemas';
 import type { BibleVersion, Collection } from './types';
 import {
   fetchFilteredCollection,
@@ -8,6 +8,9 @@ import {
   isUsableBibleVersion,
   isVersionFilterActive,
 } from './version-filters';
+
+export { GetVersionsOptionsSchema };
+export type { GetVersionsOptions };
 
 type VersionListQuery = {
   'language_ranges[]': string[];
@@ -21,32 +24,6 @@ type VersionListQuery = {
 const languageRangesSchema = z
   .string()
   .check(z.trim(), z.minLength(1, 'Language ranges must be a non-empty string'));
-
-export const GetVersionsOptionsSchema = z
-  .optional(
-    z.object({
-      page_size: z.optional(z.union([z.int().check(z.positive()), z.literal('*')])),
-      page_token: z.optional(z.string()),
-      fields: z.optional(z.array(z.keyof(BibleVersionSchema))),
-      all_available: z.optional(z.boolean()),
-    }),
-  )
-  .check(
-    z.refine(
-      (data) => {
-        if (data?.page_size === '*') {
-          return data.fields && data.fields.length >= 1 && data.fields.length <= 3;
-        }
-        return true;
-      },
-      {
-        error: 'page_size="*" required 1-3 fields to be specified',
-        path: ['page_size', 'fields'],
-      },
-    ),
-  );
-
-export type GetVersionsOptions = z.infer<typeof GetVersionsOptionsSchema>;
 
 export async function getVersions(
   client: ApiClient,
